feat: remove generic headers to simplify! part5
